Addition of BillSplitterMain
Created a new main class that also uses Calculator. This program asks the user for the bill total and number of people, and then replies with the price per person, along with any remainder. It is specifically designed to handle Japanese Yen and thus the prompts are also in Japanese.